requirements.txt
flask
requests
beautifulsoup4
openai
streamlit
sqlite3-bro
os
sqlite3
python-dotenv
firebase-admin
langchain
pymongo
scrapy
gunicorn
flask-cors